Add title screen monsters from Green and Blue as well.
Having data from all versions is a long‐term goal, but until the
ROMs for those versions can be easily built and compared to the
originals, it will be very easy for errors to slip in. Adding stuff
from Green, Blue, and Yellow is okay with me, but for now Red is
the only version guaranteed to have correct data.

hg-commit-id: c595f0981208