extract and include blockset data
extract_tileblocks also has the ability to output asm, but the output
seems like it would be too much for rgbasm to handle. Instead of
cluttering up common.asm, I am opting to place the data in binary
data files under gfx/blocksets/ in .bst files.

Every 16 bytes is a separate block in the blockset. Each byte is
referencing a tile id from the tileset graphics in gfx/tilesets/.

hg-commit-id: 0540eff0f081