text insertion code for unnamed TX_FARs
These TX_FARs are found in ASM inside INCBIN intervals, and as a
consequence do not have good names. Someone will have to review the
naming.

Note that these texts may or may not be referenced in scripts that
will eventually be imported. Some of these are raw texts that could
be completely unreferenced, but so far that doesn't look like the
case.

hg-commit-id: 47239e73071a