Restore auto-loading of default "main" asm file
This feature was removed misunderstanding its actual use: if filename
passed to load_asm() is in defaults the correct current main asm file is
loaded. This saves us from knowing which actually is the current name of
the "main" asm file, because the correct one is chosen automatically
(unless, of course, the passed filename is not in the defaults list.