Fix Cinnabar Mansion softlock
"In Generation I, it is possible to get stuck on 3F. This happens if the player stands on the east side of the table in the southwestern room and the Burglar with a Ninetales then challenges them to a battle. After the battle, the player will be stuck in a corner, the Burglar blocking the way. If the player saves after this and doesn't have an Escape Rope or a Pokémon with Dig, the only way to escape is to black out by fighting wild Pokémon. The layout was changed in FireRed and LeafGreen, and it is no longer possible to get trapped."

Fixed this by changing the rock formations, eliminates all possibilities.