Revamp Pokemon Mansion B2F
This was hell to do. There is still one bug in there, and that's the item pickups don't work for some reason. I don't know how these work honestly so if anyone wants to sort that go ahead.

Anyway, some stuff on the map;
- Currently, a warp from B1F leads there, which is intentionally non-returnable. The player walks through the whole map to get to the warp panel. A walk of humanity's shame, if you will. A warp for where the player is meant to go will be necessary later.
- Lore from Japanese media, particularly the first movie's interpretation of Mewtwo, is included. This is based on context clues on its canonicity from LGPE and other bits.
- A clue on how to get the Old Sea Chart is not-so-subtly included.

The MissingNo. event is unfinished, I intend for Martha to deal with that. The text intentionally lacks a "done" bit to force an unpredictable error; I believe this is inconsequential gameplay-wise and should make for a good opening quote when fighting it.

Oh, and Faraway Island has Mankey, Butterfree, and Primeape as find-able Pokemon now. This should be expanded upon later.