Faster health bar
A simple 1-character change that does so much. Usually, the game delays it by 2 frames per pixel, so changing it to 1 doubles the speed. If it remains too slow, I believe it can be removed entirely, but I'm avoiding it in case of bugs.

I also changed the debug Pokemon to Croakozuna as Pound is s/e vs Dark, allowing for an easier visual. Rhydon changed to Rhyperior for the same reason; more damage, better visual.