Faraway Island warp now works
Essentially, I bruteforced making it work; made the tiles we desire be warps for us instead. You do this by noting the bottom-left tile of the block for some reason, so tile $30 of forest will be a warp tile even though it really looks like it shouldn't be. As far as I can tell, there are no consequences for this war crime code as long as you don't put a warp where it shouldn't be.