Up-Grade NPC
Seems to interact oddly with Saffron but I believe this is because of the way Team Rocket works. The event itself works fine.

Also added the Up-Grade, as before, Porygon was evolving via Trade.

Uses text referencing LGPE, and has a similar location.

Because of the sprite limit in Saffron, the Pidgeot NPC has been removed.