Groundwork for Viridian Pre-Gym
This map is fully functional but has a lot of bits to iron out, as well as a need for Gym Trainers. I'm also not wholly satisfied with Yujirou's team; maybe remove Eevee?

Yes, I designed an entire tileset for this. Fear me. Bow before me. Actually, don't, this was beyond rational thought.

Current issues;
- Yujirou's sprite doesn't work properly.
- Needs Gym Trainers (hell)
- Yujirou needs an event for being beaten, the refight, and more.