Extra trainer classes, new bold P
I changed the P image to match the font used in the prototype assets.

Also, because Yujirou has a pic all alone in that one bank, I added the other 4 beta trainer classes that never got used. They're all ready to be used, just need teams and a place to go, but I want to address trainer line-ups last code-wise.