Refactor Item Constants
Alright, never put items after the floors. This is very, very bad, as it actually causes a myriad of problems whenever the game tries to assert the item list on its lonesome. Effectively, this has been adding floors. This caused the item descriptions to be offset by 21 spaces once you get to the TMs. The fact it worked before was a miracle.

Anyway, all considerations have been made when doing this, so it should all be just fine.

If you need to test if the items are ok, check the Up-Grade on Porygon, as it's the last in the list. Checks showed it was ok for me.

Also, Old Sea Chart now uses the localised name - Old Sea Map. This actually fits in-game, which is good.