Finish Rocket House, Yellow sprite backport
Some big dubs for KEP today. This commit backports the Yellow overworld spriteset, and also adds the cat sprite for Meowth. This allows Jessie and James to use their overworld sprites in the Rocket House and serves as groundwork for backporting their other events.

There are altered Daisy and Nurse Joy sprites that may or may not warrant usage. That can be discussed.

Naturally, this also means there's a ton of overworld sprites to use elsewhere. Revamps of the "Bulbasaur house" in Cerulean and the Surfing Pikachu sprite change come to mind.

An & symbol has been backported from GSC for Jessie and James's trainer class as well.