KEP tweaks post-playtest
This fixes the rival team selection issue and changes the wild data of Route 25.

The rival pulls data using hex and I put in dec like an idiot. Also, fixed the ordering.

Paras and Skimper make little sense on Route 25 and felt honestly quite immersion-breaking. Paras is a cave mushroom crab thing, it wouldn't be seen on a seaside route like Route 25. Skimper just didn't feel right on a similar basis - just...why would it be there? It's a newt, right? Instead, I added Venonat, a Pokemon seen in Yellow, which needs to be here to benefit from the learnset tweaks that we backported.