Populate Citrine City
This commit adds NPCs and signs to Citrine, doing some significant worldbuilding in the process. I considered adding an event where Oak greets you like how the Magma/Aqua Admins do on the ORAS Battle Resort, but it seemed a little too on-the-nose, y'know?

There was an issue where you wouldn't be walked out of doors automatically, which was quite immersion-breaking. That was because the door tile IDs weren't defined - this is fixed too.

Some map elements have been altered to facilitate the new changes, and obviously, some map sprite slots have been filled.

Importing and changing the Battle Tent is most likely my next task.