DV / Stat Exp display in stat menu
Imported from Shin Pokered. Required a little bit of refactoring but all in all very straightforward. I changed the WRAM associations to be easier for me to work with should I ever need to.

I am not a 100% fan of the way this works - requiring hold inputs - but I am also not sure how to make it just be something you can easily input...