Correct wTilePlayerStandingOn "unused?" comment (#433)
`TilePairCollisionsWater` contains the entry `db CAVERN, $14, $05`, which prevents surfing from raised cave ground tiles. If `wTilePlayerStandingOn` is not updated when the start menu is opened, it will be left at the previous tile, which will be the raised ground tile if you have just walked onto the stairs from one.