Fix stuck invulnerable glitch (original game)
Clear the invulnerability flag when you get fully paralyzed or confused while using dig or fly in all scenarios EXCEPT if it happens to the player in normal play against NPCs/wild pokemon. Because in that scenario it's just amusing. In link battles it gets cleared and if it happens to the opponent it gets cleared.