Correct wTileMapBackup declared space
wTileMapBackup isnt just used to backup the tilemap but also as a buffer for a screen view out of blocks during the drawing of the tilemap, define a new union reflecting that.
Add map gfx values in gfx_constants.asm.
Edit overworld.asm to replace hardcoded values with new constants.